change some code for valdation

+++ app/workers/delete_post_worker.rb
@@ -7,7 +7,8 @@ class DeletePostWorker
 
   def perform(id)
     post = Post.find_by_id(id)
-    &. post
-    post.destroy
+    if post
+      post.destroy
+    end
   end
 end
